Rename temp log service -> coldchain
