+++ server/graphql/cold_chain/src/types/sensor.rs
@@ -11,8 +11,8 @@ use repository::{
     TemperatureLogFilter, TemperatureLogSort, TemperatureLogSortField,
 };
 use repository::{EqualFilter, Sensor, SensorFilter, SensorRow, SensorSort, SensorSortField};
+use service::cold_chain::query_temperature_log::get_temperature_logs;
 use service::temperature_breach::query::temperature_breaches;
-use service::temperature_log::query::get_temperature_logs;
 use service::{usize_to_u32, ListResult};
 
 use super::temperature_breach::TemperatureBreachNodeType;
