+++ server/service/src/cold_chain/mod.rs
@@ -1,15 +1,19 @@
-use self::insert::{insert_temperature_log, InsertTemperatureLog, InsertTemperatureLogError};
-use self::query::{get_temperature_log, get_temperature_logs};
-use self::update::{update_temperature_log, UpdateTemperatureLog, UpdateTemperatureLogError};
+use self::insert_temperature_log::{
+    insert_temperature_log, InsertTemperatureLog, InsertTemperatureLogError,
+};
+use self::query_temperature_log::{get_temperature_log, get_temperature_logs};
+use self::update_temperature_log::{
+    update_temperature_log, UpdateTemperatureLog, UpdateTemperatureLogError,
+};
 
 use super::{ListError, ListResult};
 use crate::{service_provider::ServiceContext, SingleRecordError};
 use repository::temperature_log::{TemperatureLog, TemperatureLogFilter, TemperatureLogSort};
 use repository::{PaginationOption, StorageConnection};
 
-pub mod insert;
-pub mod query;
-pub mod update;
+pub mod insert_temperature_log;
+pub mod query_temperature_log;
+pub mod update_temperature_log;
 mod validate;
 
 pub trait TemperatureLogServiceTrait: Sync + Send {
