+++ server/service/src/cold_chain/tests/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(test)]
+mod query_temperature_log;
