+++ server/service/src/cold_chain/validate.rs
