+++ server/service/src/lib.rs
@@ -16,6 +16,7 @@ pub mod auth_data;
 pub mod barcode;
 pub mod catalogue;
 pub mod clinician;
+pub mod cold_chain;
 mod common_stock;
 pub mod currency;
 pub mod cursor_controller;
@@ -59,7 +60,6 @@ pub mod system_user;
 pub mod temperature_breach;
 pub mod temperature_chart;
 pub mod temperature_excursion;
-pub mod temperature_log;
 pub mod token;
 pub mod token_bucket;
 pub mod user_account;
