--- server/service/src/temperature_log/tests/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(test)]
-mod query;
